requirements-dev.txt
appdirs==1.4.4
astroid==2.15.1
attrs==22.2.0
black==23.1.0
certifi==2022.12.7
charset-normalizer==2.0.10
click==8.1.3
colorama==0.4.6
Django==4.2
django-allauth==0.54.0
django-ninja==0.21
django-stubs[compatible-mypy]==4.2.0
dropbox==11.36.0
dj-database-url==1.3.0
ghp-import==2.1.0
idna==3.3
importlib-metadata==4.10.0
isort==5.10.1
Jinja2==3.0.3
lazy-object-proxy==1.9.0
Markdown==3.3.6
MarkupSafe==2.1.2
mccabe==0.7.0
mergedeep==1.3.4
mkdocs==1.4.2
mkdocs-material==9.1.5
mkdocs-material-extensions==1.1.1
mkdocs-jupyter==0.24.1
mypy==1.2.0
packaging==22.0
pathspec==0.11.0
pylint==2.17.1
pylint-django==2.5.3
platformdirs==2.4.1
pymdown-extensions==9.10
pyparsing==3.0.6
pyrsistent==0.17.3
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
pyyaml_env_tag==0.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.8
watchdog==2.1.6
whitenoise==5.2.0
wincertstore==0.2
zipp==3.7.0
